Pavel/увеид tree list little adj (#1482)
* put checkbox back to radioGroups select

* table header cell to allow 2 lines

* rte - clean pasted styles

* bootstrap

* link clicked event

* icon rotate input

* fixed tests

* smooth scroll fix

* misc upd

* fixed test

* ListPanelService, new icon+badge

* a little better

* misc

* fixed test

* selectMode fix

* Merge remote-tracking branch 'origin/master' into pavel/mar15-misc

* wip

* bottom-right badge position for online/offile

* same

* wip

* menu upd

* same

* same

* stash

* and some more

* some more

* fixed test

* same

* forgot smth

* still menu

* wip

* fix

* editable tree list - part 1

* same

* public api

* update storybook & tsconfig

* tab action

* visual options

* collapsible

* same

* a little better collapse

* emit on text change

* tree list editor update

* fixed type

* wip

* basic d-n-d styling

* working on DnD

* fixed test

* little better

* some more better

* chip-input upd

* little cleanup

* little upd

* WIP

* WIP

* drop seems to work

* even better

* more better

* some refactoring

* drag cancel & group open on drag hover

* little things

* safer, better

* same

* video embed - vimeo thumbnail fix

* more refactoriing

* tree panel width 500

* editable tree list - focus input

* same

* some more

* some safeguarding

* same

* more focus-related

+++ projects/ui-framework/src/lib/lists/tree-list/editable-tree-list/editable-tree-list.abstract.ts
@@ -124,6 +124,7 @@ export abstract class BaseEditableTreeListElement
   protected dragHoverTimer;
   protected dragRef: DragRef<any>;
   protected cancelDrop = false;
+  protected cancelFocus = false;
   protected expandedWhileDragging: Set<TreeListItem> = new Set();
   private windowKeydownSubscriber: Subscription;
 
@@ -136,6 +137,10 @@ export abstract class BaseEditableTreeListElement
   protected listBackup: TreeListOption[];
 
   @HostListener('click', ['$event']) onHostClick(event: MouseEvent) {
+    if (this.cancelFocus) {
+      this.cancelFocus = false;
+      return;
+    }
     const listLength = this.listViewModel?.length || 0;
     this.focus(
       event.target !== this.hostElement || listLength < 2
@@ -305,7 +310,8 @@ export abstract class BaseEditableTreeListElement
   }
 
   public onListClick(event: MouseEvent): void {
-    const clickedElement: HTMLElement = this.cntrlsSrvc.onListClick(event, {
+    this.cancelFocus = false;
+    this.cntrlsSrvc.onListClick(event, {
       itemsMap: this.itemsMap,
       listViewModel: this.listViewModel,
       toggleItemCollapsed: this.toggleItemCollapsed.bind(this),
@@ -436,11 +442,18 @@ export abstract class BaseEditableTreeListElement
           this.itemsMap,
           this.listViewModel
         );
-        if (item && !item.childrenCount) {
+        if (item && !item.childrenCount && this.listViewModel.length > 1) {
           this.deleteItem(item);
         }
+        this.cancelFocus = true;
       }
     }
+
+    this.zone.runOutsideAngular(() => {
+      setTimeout(() => {
+        this.cancelFocus = false;
+      }, 100);
+    });
   }
 
   public focus(whichInput: 'first' | 'last' | number = 'first'): void {
